+++ src/client.h
@@ -33,6 +33,6 @@ pthread_t *launch_threads();
 void process_stats();
 void run_iterations();
 void cleanup();
-void write_logFile(const char * type, int size, int duration, int start_time);
+void write_logFile(const char * type, int size, int duration, long long start_time);
 
 #endif
